ISSUE-1024 Refactor TopologyCatalogResource (also API endpoint paths)
* split up TopologyCatalogResource into three pieces
  * new resources: TopologyDashboardResource, TopologyActionResource
* change some of REST API endpoints which are moved to TopologyDashboardResource
* modify UI side as well
