+++ streams/service/src/main/java/com/hortonworks/streamline/streams/service/TopologyCatalogResource.java
@@ -18,24 +18,16 @@ package com.hortonworks.streamline.streams.service;
 
 import com.codahale.metrics.annotation.Timed;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.base.Stopwatch;
-import com.hortonworks.registries.common.transaction.TransactionIsolation;
-import com.hortonworks.registries.storage.TransactionManager;
 import com.hortonworks.streamline.common.exception.service.exception.request.BadRequestException;
 import com.hortonworks.streamline.common.exception.service.exception.request.EntityNotFoundException;
-import com.hortonworks.streamline.common.exception.service.exception.request.TopologyAlreadyExistsOnCluster;
 import com.hortonworks.streamline.common.exception.service.exception.server.StreamingEngineNotReachableException;
-import com.hortonworks.streamline.common.util.ParallelStreamUtil;
 import com.hortonworks.streamline.common.util.WSUtils;
-import com.hortonworks.streamline.streams.actions.TopologyActions;
 import com.hortonworks.streamline.streams.actions.topology.service.TopologyActionsService;
 import com.hortonworks.streamline.streams.catalog.Topology;
 import com.hortonworks.streamline.streams.catalog.TopologyVersion;
 import com.hortonworks.streamline.streams.catalog.service.StreamCatalogService;
 import com.hortonworks.streamline.streams.catalog.topology.TopologyData;
-import com.hortonworks.streamline.streams.cluster.service.EnvironmentService;
 import com.hortonworks.streamline.streams.exception.TopologyNotAliveException;
-import com.hortonworks.streamline.streams.metrics.topology.service.TopologyMetricsService;
 import com.hortonworks.streamline.streams.security.Permission;
 import com.hortonworks.streamline.streams.security.Roles;
 import com.hortonworks.streamline.streams.security.SecurityUtil;
@@ -62,15 +54,11 @@ import javax.ws.rs.core.SecurityContext;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.EnumSet;
-import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.ForkJoinPool;
-import java.util.concurrent.TimeUnit;
 
 import static com.hortonworks.streamline.streams.catalog.Topology.NAMESPACE;
 import static com.hortonworks.streamline.streams.catalog.TopologyVersion.VERSION_PREFIX;
@@ -78,8 +66,6 @@ import static com.hortonworks.streamline.streams.security.Permission.DELETE;
 import static com.hortonworks.streamline.streams.security.Permission.EXECUTE;
 import static com.hortonworks.streamline.streams.security.Permission.READ;
 import static com.hortonworks.streamline.streams.security.Permission.WRITE;
-import static com.hortonworks.streamline.streams.service.TopologySortType.LAST_UPDATED;
-import static java.util.stream.Collectors.toList;
 import static javax.ws.rs.core.Response.Status.CREATED;
 import static javax.ws.rs.core.Response.Status.OK;
 
@@ -88,42 +74,22 @@ import static javax.ws.rs.core.Response.Status.OK;
 @Produces(MediaType.APPLICATION_JSON)
 public class TopologyCatalogResource {
     private static final Logger LOG = LoggerFactory.getLogger(TopologyCatalogResource.class);
-    private static final Integer DEFAULT_N_OF_TOP_N_LATENCY = 3;
-    private static final String DEFAULT_SORT_TYPE = LAST_UPDATED.name();
-    private static final Boolean DEFAULT_SORT_ORDER_ASCENDING = false;
-
-    private static final int FORK_JOIN_POOL_PARALLELISM = 10;
-    private final ForkJoinPool forkJoinPool = new ForkJoinPool(FORK_JOIN_POOL_PARALLELISM);
-
-    private final URL SCHEMA = Thread.currentThread().getContextClassLoader()
-            .getResource("assets/schemas/topology.json");
 
     private final StreamlineAuthorizer authorizer;
     private final StreamCatalogService catalogService;
-    private final EnvironmentService environmentService;
     private final TopologyActionsService actionsService;
-    private final TopologyMetricsService metricsService;
-    private final TransactionManager transactionManager;
 
     public TopologyCatalogResource(StreamlineAuthorizer authorizer, StreamCatalogService catalogService,
-                                   EnvironmentService environmentService, TopologyActionsService actionsService,
-                                   TopologyMetricsService metricsService, TransactionManager transactionManager) {
+                                   TopologyActionsService actionsService) {
         this.authorizer = authorizer;
         this.catalogService = catalogService;
-        this.environmentService = environmentService;
         this.actionsService = actionsService;
-        this.metricsService = metricsService;
-        this.transactionManager = transactionManager;
     }
 
     @GET
     @Path("/topologies")
     @Timed
-    public Response listTopologies (@javax.ws.rs.QueryParam("detail") Boolean detail,
-                                    @javax.ws.rs.QueryParam("sort") String sortType,
-                                    @javax.ws.rs.QueryParam("ascending") Boolean ascending,
-                                    @javax.ws.rs.QueryParam("latencyTopN") Integer latencyTopN,
-                                    @Context SecurityContext securityContext) {
+    public Response listTopologies (@Context SecurityContext securityContext) {
         Collection<Topology> topologies = catalogService.listTopologies();
         boolean topologyUser = SecurityUtil.hasRole(authorizer, securityContext, Roles.ROLE_TOPOLOGY_USER);
         if (topologyUser) {
@@ -133,22 +99,7 @@ public class TopologyCatalogResource {
         }
         Response response;
         if (topologies != null) {
-            if (detail == null || !detail) {
-                response = WSUtils.respondEntities(topologies, OK);
-            } else {
-                if (sortType == null) {
-                    sortType = DEFAULT_SORT_TYPE;
-                }
-                if (ascending == null) {
-                    ascending = DEFAULT_SORT_ORDER_ASCENDING;
-                }
-
-                String asUser = WSUtils.getUserFromSecurityContext(securityContext);
-                List<CatalogResourceUtil.TopologyDetailedResponse> detailedTopologies = enrichTopologies(topologies, asUser,
-                        sortType, ascending, latencyTopN);
-
-                response = WSUtils.respondEntities(detailedTopologies, OK);
-            }
+            response = WSUtils.respondEntities(topologies, OK);
         } else {
             response = WSUtils.respondEntities(Collections.emptyList(), OK);
         }
@@ -160,17 +111,13 @@ public class TopologyCatalogResource {
     @Path("/topologies/{topologyId}")
     @Timed
     public Response getTopologyById(@PathParam("topologyId") Long topologyId,
-                                    @javax.ws.rs.QueryParam("detail") Boolean detail,
-                                    @javax.ws.rs.QueryParam("latencyTopN") Integer latencyTopN,
                                     @Context SecurityContext securityContext) {
         SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_USER,
                 NAMESPACE, topologyId, READ);
-        String asUser = WSUtils.getUserFromSecurityContext(securityContext);
-        Response response = getTopologyByIdAndVersionId(topologyId,
-                catalogService.getCurrentVersionId(topologyId),
-                detail, latencyTopN, asUser);
-        if (response != null) {
-            return response;
+
+        Topology result = catalogService.getTopology(topologyId);
+        if (result != null) {
+            return WSUtils.respondEntity(result, OK);
         }
 
         throw EntityNotFoundException.byId(topologyId.toString());
@@ -181,34 +128,16 @@ public class TopologyCatalogResource {
     @Timed
     public Response getTopologyByIdAndVersion(@PathParam("topologyId") Long topologyId,
                                               @PathParam("versionId") Long versionId,
-                                              @javax.ws.rs.QueryParam("detail") Boolean detail,
-                                              @javax.ws.rs.QueryParam("latencyTopN") Integer latencyTopN,
                                               @Context SecurityContext securityContext) {
         SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_USER,
                 NAMESPACE, topologyId, READ);
-        String asUser = WSUtils.getUserFromSecurityContext(securityContext);
-        Response response = getTopologyByIdAndVersionId(topologyId, versionId, detail, latencyTopN, asUser);
-        if (response != null) {
-            return response;
-        }
 
-        throw EntityNotFoundException.byVersion(topologyId.toString(), versionId.toString());
-    }
-
-    private Response getTopologyByIdAndVersionId(Long topologyId, Long versionId, Boolean detail,
-                                                 Integer latencyTopN, String asUser) {
         Topology result = catalogService.getTopology(topologyId, versionId);
         if (result != null) {
-            if (detail == null || !detail) {
-                return WSUtils.respondEntity(result, OK);
-            } else {
-                CatalogResourceUtil.TopologyDetailedResponse topologyDetailed =
-                        CatalogResourceUtil.enrichTopology(result, asUser, latencyTopN,
-                                environmentService, actionsService, metricsService, catalogService);
-                return WSUtils.respondEntity(topologyDetailed, OK);
-            }
+            return WSUtils.respondEntity(result, OK);
         }
-        return null;
+
+        throw EntityNotFoundException.byId(topologyId.toString());
     }
 
     @GET
@@ -227,7 +156,6 @@ public class TopologyCatalogResource {
         }
         return response;
     }
-
 
     @POST
     @Path("/topologies")
@@ -410,7 +338,6 @@ public class TopologyCatalogResource {
         throw EntityNotFoundException.byVersion(topologyId.toString(), versionId.toString());
     }
 
-
     @GET
     @Path("/topologies/{topologyId}/deploymentstate")
     @Timed
@@ -419,208 +346,6 @@ public class TopologyCatalogResource {
                 .flatMap(t -> catalogService.getTopologyState(t.getId()))
                 .map(s -> WSUtils.respondEntity(s, OK))
                 .orElseThrow(() -> EntityNotFoundException.byId(topologyId.toString()));
-    }
-
-    @GET
-    @Path("/topologies/{topologyId}/actions/status")
-    @Timed
-    public Response topologyStatus (@PathParam("topologyId") Long topologyId,
-                                    @Context SecurityContext securityContext) throws Exception {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology result = catalogService.getTopology(topologyId);
-        if (result != null) {
-            String asUser = WSUtils.getUserFromSecurityContext(securityContext);
-            TopologyActions.Status status = actionsService.topologyStatus(result, asUser);
-            return WSUtils.respondEntity(status, OK);
-        }
-
-        throw EntityNotFoundException.byId(topologyId.toString());
-    }
-
-    @GET
-    @Path("/topologies/{topologyId}/versions/{versionId}/actions/status")
-    @Timed
-    public Response topologyStatusVersion(@PathParam("topologyId") Long topologyId,
-                                          @PathParam("versionId") Long versionId,
-                                          @Context SecurityContext securityContext) throws Exception {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology result = catalogService.getTopology(topologyId, versionId);
-        if (result != null) {
-            String asUser = WSUtils.getUserFromSecurityContext(securityContext);
-            TopologyActions.Status status = actionsService.topologyStatus(result, asUser);
-            return WSUtils.respondEntity(status, OK);
-        }
-
-        throw EntityNotFoundException.byVersion(topologyId.toString(), versionId.toString());
-    }
-
-    @POST
-    @Path("/topologies/{topologyId}/actions/validate")
-    @Timed
-    public Response validateTopology (@PathParam("topologyId") Long topologyId, @Context SecurityContext securityContext) {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology result = catalogService.getTopology(topologyId);
-        if (result != null) {
-            //catalogService.validateTopology(SCHEMA, topologyId);
-            return WSUtils.respondEntity(result, OK);
-        }
-
-        throw EntityNotFoundException.byId(topologyId.toString());
-    }
-
-    @POST
-    @Path("/topologies/{topologyId}/versions/{versionId}/actions/validate")
-    @Timed
-    public Response validateTopologyVersion(@PathParam("topologyId") Long topologyId,
-                                            @PathParam("versionId") Long versionId, @Context SecurityContext securityContext) {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology result = catalogService.getTopology(topologyId, versionId);
-        if (result != null) {
-            //catalogService.validateTopology(SCHEMA, topologyId);
-            return WSUtils.respondEntity(result, OK);
-        }
-
-        throw EntityNotFoundException.byVersion(topologyId.toString(), versionId.toString());
-    }
-
-    @POST
-    @Path("/topologies/{topologyId}/actions/deploy")
-    @Timed
-    public Response deployTopology (@PathParam("topologyId") Long topologyId, @Context SecurityContext securityContext) throws Exception {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology topology = catalogService.getTopology(topologyId);
-        if (topology != null) {
-            return deploy(topology, securityContext);
-        }
-        throw EntityNotFoundException.byId(topologyId.toString());
-    }
-
-
-    @POST
-    @Path("/topologies/{topologyId}/versions/{versionId}/actions/deploy")
-    @Timed
-    public Response deployTopologyVersion(@PathParam("topologyId") Long topologyId,
-                                          @PathParam("versionId") Long versionId,
-                                          @Context SecurityContext securityContext) throws Exception {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology topology = catalogService.getTopology(topologyId, versionId);
-        if (topology != null) {
-            return deploy(topology, securityContext);
-        }
-        throw EntityNotFoundException.byVersion(topologyId.toString(), versionId.toString());
-    }
-
-    private Response deploy(Topology topology, SecurityContext securityContext) {
-        String asUser = WSUtils.getUserFromSecurityContext(securityContext);
-        try {
-            ParallelStreamUtil.runAsync(() -> actionsService.deployTopology(topology, asUser), forkJoinPool);
-            return WSUtils.respondEntity(topology, OK);
-        } catch (TopologyAlreadyExistsOnCluster ex) {
-            return ex.getResponse();
-        }
-    }
-
-    @POST
-    @Path("/topologies/{topologyId}/actions/kill")
-    @Timed
-    public Response killTopology (@PathParam("topologyId") Long topologyId, @Context SecurityContext securityContext) throws Exception {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology result = catalogService.getTopology(topologyId);
-        if (result != null) {
-            actionsService.killTopology(result, WSUtils.getUserFromSecurityContext(securityContext));
-            return WSUtils.respondEntity(result, OK);
-        }
-
-        throw EntityNotFoundException.byId(topologyId.toString());
-    }
-
-    @POST
-    @Path("/topologies/{topologyId}/versions/{versionId}/actions/kill")
-    @Timed
-    public Response killTopologyVersion(@PathParam("topologyId") Long topologyId,
-                                        @PathParam("versionId") Long versionId,
-                                        @Context SecurityContext securityContext) throws Exception {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology result = catalogService.getTopology(topologyId, versionId);
-        if (result != null) {
-            actionsService.killTopology(result, WSUtils.getUserFromSecurityContext(securityContext));
-            return WSUtils.respondEntity(result, OK);
-        }
-
-        throw EntityNotFoundException.byVersion(topologyId.toString(), versionId.toString());
-    }
-
-    @POST
-    @Path("/topologies/{topologyId}/actions/suspend")
-    @Timed
-    public Response suspendTopology (@PathParam("topologyId") Long topologyId, @Context SecurityContext securityContext) throws Exception {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology result = catalogService.getTopology(topologyId);
-        if (result != null) {
-            actionsService.suspendTopology(result, WSUtils.getUserFromSecurityContext(securityContext));
-            return WSUtils.respondEntity(result, OK);
-        }
-
-        throw EntityNotFoundException.byId(topologyId.toString());
-    }
-
-    @POST
-    @Path("/topologies/{topologyId}/versions/{versionId}/actions/suspend")
-    @Timed
-    public Response suspendTopologyVersion(@PathParam("topologyId") Long topologyId,
-                                           @PathParam("versionId") Long versionId,
-                                           @Context SecurityContext securityContext) throws Exception {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology result = catalogService.getTopology(topologyId, versionId);
-        if (result != null) {
-            actionsService.suspendTopology(result, WSUtils.getUserFromSecurityContext(securityContext));
-            return WSUtils.respondEntity(result, OK);
-        }
-
-        throw EntityNotFoundException.byVersion(topologyId.toString(), versionId.toString());
-    }
-
-    @POST
-    @Path("/topologies/{topologyId}/actions/resume")
-    @Timed
-    public Response resumeTopology (@PathParam("topologyId") Long topologyId,
-                                    @Context SecurityContext securityContext) throws Exception {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology result = catalogService.getTopology(topologyId);
-        if (result != null) {
-            actionsService.resumeTopology(result, WSUtils.getUserFromSecurityContext(securityContext));
-            return WSUtils.respondEntity(result, OK);
-        }
-
-        throw EntityNotFoundException.byId(topologyId.toString());
-    }
-
-    @POST
-    @Path("/topologies/{topologyId}/versions/{versionId}/actions/resume")
-    @Timed
-    public Response resumeTopologyVersion(@PathParam("topologyId") Long topologyId,
-                                          @PathParam("versionId") Long versionId,
-                                          @Context SecurityContext securityContext) throws Exception {
-        SecurityUtil.checkRoleOrPermissions(authorizer, securityContext, Roles.ROLE_TOPOLOGY_SUPER_ADMIN,
-                NAMESPACE, topologyId, READ, EXECUTE);
-        Topology result = catalogService.getTopology(topologyId, versionId);
-        if (result != null) {
-            actionsService.resumeTopology(result, WSUtils.getUserFromSecurityContext(securityContext));
-            return WSUtils.respondEntity(result, OK);
-        }
-
-        throw EntityNotFoundException.byVersion(topologyId.toString(), versionId.toString());
     }
 
     @GET
@@ -644,7 +369,6 @@ public class TopologyCatalogResource {
 
         throw EntityNotFoundException.byId(topologyId.toString());
     }
-
 
     @POST
     @Path("/topologies/{topologyId}/actions/clone")
@@ -701,52 +425,5 @@ public class TopologyCatalogResource {
         throw EntityNotFoundException.byId(topologyId.toString());
     }
 
-    private List<CatalogResourceUtil.TopologyDetailedResponse> enrichTopologies(Collection<Topology> topologies, String asUser, String sortType, Boolean ascending, Integer latencyTopN) {
-        LOG.debug("[START] enrichTopologies");
-        Stopwatch stopwatch = Stopwatch.createStarted();
-
-        try {
-            List<CatalogResourceUtil.TopologyDetailedResponse> responses = ParallelStreamUtil.execute(() ->
-                    topologies.parallelStream()
-                            .map(t -> {
-                                try {
-                                    transactionManager.beginTransaction(TransactionIsolation.DEFAULT);
-                                    CatalogResourceUtil.TopologyDetailedResponse topologyDetailedResponse = CatalogResourceUtil.enrichTopology(t, asUser, latencyTopN,
-                                            environmentService, actionsService, metricsService, catalogService);
-                                    transactionManager.commitTransaction();
-                                    return topologyDetailedResponse;
-                                } catch (Exception e) {
-                                    transactionManager.rollbackTransaction();
-                                    throw e;
-                                }
-                            })
-                            .sorted((c1, c2) -> {
-                                int compared;
-
-                                switch (TopologySortType.valueOf(sortType.toUpperCase())) {
-                                    case NAME:
-                                        compared = c1.getTopology().getName().compareTo(c2.getTopology().getName());
-                                        break;
-                                    case STATUS:
-                                        compared = c1.getRunning().compareTo(c2.getRunning());
-                                        break;
-                                    case LAST_UPDATED:
-                                        compared = c1.getTopology().getVersionTimestamp().compareTo(c2.getTopology().getVersionTimestamp());
-                                        break;
-                                    default:
-                                        throw new IllegalStateException("Not supported SortType: " + sortType);
-                                }
-
-                                return ascending ? compared : (compared * -1);
-                            })
-                            .collect(toList()), forkJoinPool);
-
-            LOG.debug("[END] enrichTopologies - elapsed: {} ms", stopwatch.elapsed(TimeUnit.MILLISECONDS));
-
-            return responses;
-        } finally {
-            stopwatch.stop();
-        }
-    }
 }
